feat: 디자인 시스템/ 커스텀 바텀시트 추가 (isModalInPresentation 대응)

+++ DesignSystem/Sources/Foundation/Shadows/Shadow.swift
@@ -9,7 +9,7 @@ import UIKit
 
 public struct Shadow: Sendable {
     public let shadowColor: UIColor
-    public let shadowOpacity: CGFloat
+    public let shadowOpacity: Float
     public let shadowOffset: CGSize
     public let shadowRadius: CGFloat
 }
